David, Hari | #1962 | Implemented logic to clear data sync failure after merging module data block

+++ src/main/js/app/consumer/mergestrategies/module.data.block.merger.js
@@ -88,7 +88,14 @@ define(['moment', 'lodash'],
                 };
 
                 var resetDataSyncFailure = function () {
-                    return dataSyncFailureRepository.delete(moduleDataBlock.moduleId, moduleDataBlock.period);
+                    var mergedDataIsEqualToPraxisDataButNotDhis = mergedDataValuesAreEqualToExistingPraxisDataValues() && !mergedDataValuesAreEqualToDhisDataValues(),
+                        approvedAtProjectLevelOnlyOnPraxis = moduleDataBlock.approvedAtProjectLevel && !dhisCompletion,
+                        approvedAtCoordinationLevelOnlyOnPraxis = moduleDataBlock.approvedAtCoordinationLevel && !dhisApproval;
+
+                    if(moduleDataBlock.failedToSync && !mergedDataIsEqualToPraxisDataButNotDhis &&
+                       !approvedAtProjectLevelOnlyOnPraxis && !approvedAtCoordinationLevelOnlyOnPraxis) {
+                        return dataSyncFailureRepository.delete(moduleDataBlock.moduleId, moduleDataBlock.period);
+                    }
                 };
 
                 return mergeAndSaveDataValues()
